mysite/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==2.1
django-ckeditor==5.9.0
django-js-asset==1.2.3
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
